+++ nuxt.config.js
@@ -22,12 +22,17 @@ export default {
   buildModules: [
     '@nuxtjs/eslint-module',
     '@nuxtjs/vuetify',
-    ['@pinia/nuxt', { disableVuex: false}]
+    //'@nuxtjs/composition-api/module', //TODO: Chokes on babel.config.js if uncommented.
+    ['@pinia/nuxt', { disableVuex: true}]
   ],
   modules: ['@nuxtjs/axios', '@nuxtjs/pwa', ['nuxt-highcharts', {}]],
   axios: { baseURL: '/', headers: { common: { Accept: 'application/json' }}},
   pwa: { manifest: { lang: 'en' }},
   vuetify: { defaultAssets: { icons: 'fa' }},
+  // Added for nuxt 4 compliance
+  pinia: {
+    storesDirs: ['./stores/**', './src/stores/**']
+  },
   static: { prefix: false },
   srcDir: 'src/',
   router: { middleware: 'auth' },
