+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@nuxtjs/axios": "^5.13.6",
     "@nuxtjs/pwa": "^3.3.5",
+    "@pinia/testing": "^0.1.6",
     "animate.css": "^3.7.0",
     "axios": "^0.21.4",
     "core-js": "^3.19.3",
@@ -28,16 +29,15 @@
     "vue-template-compiler": "2.7.14",
     "vue2-particles": "^2.9.3",
     "vuetify": "^2.6.1",
-    "vuex": "3.6.2",
     "webpack": "^4.46.0",
     "yaml-loader": "^0.6.0"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.16.3",
+    "@nuxtjs/composition-api": "0.29.3",
     "@nuxtjs/eslint-config": "^8.0.0",
     "@nuxtjs/eslint-module": "^3.0.2",
     "@nuxtjs/vuetify": "^1.12.3",
-    "@nuxtjs/composition-api": "0.29.3",
     "@pinia/nuxt": "0.2.1",
     "@vue/cli-plugin-babel": "5.0.8",
     "@vue/test-utils": "^1.3.0",
