+++ src/components/creator/exposure-information/chemicals-autocomplete.vue
@@ -96,9 +96,11 @@
 
 <script>
 import { easeInOutQuint } from 'vuetify/lib/services/goto/easing-patterns'
-import { mapGetters, mapActions, mapMutations } from "vuex"
 import { searchChemicals } from "@/utils/search"
+import { useCreatorChemicalsStore } from '@/stores/creator-chemicals'
 import doseMixin from "@/mixins/doseCSSMixin"
+
+const creatorChemicals = useCreatorChemicalsStore();
 
 export default {
   name: "ChemicalAutocomplete",
@@ -112,15 +114,18 @@ export default {
   },
   computed: {
     chemicals: {
-      get() { return this.getSelectedChemicals()(this.index) || [] },
-      set(value) { this.setChemicalGroupChemicals({ index: this.index, chemicals: value }) }
+      get() { return creatorChemicals.getSelectedChemicals()(this.index) || [] },
+      set(value) { creatorChemicals.setChemicalGroupChemicals({ index: this.index, chemicals: value }) }
     }
   },
   methods: {
-    ...mapGetters("creator-chemicals", ["getAvailableChemicals", "getSelectedChemicals", "getChemical"]),
-    ...mapActions("creator-chemicals", ["setChemicalGroupChemicals"]),
-    ...mapMutations("creator-chemicals", ["removeChemicalFromSelectedGroup"]),
-    searchChemicals: (chemical, queryText) => searchChemicals(chemical, queryText)
+    searchChemicals: (chemical, queryText) => searchChemicals(chemical, queryText),
+    getAvailableChemicals: (index) => { creatorChemicals.getAvailableChemicals(index) },
+    getSelectedChemicals: (index) => { creatorChemicals.getSelectedChemicals(index) },
+    getChemical: (name) => { creatorChemicals.getChemical(name) },
+    removeChemicalFromSelectedGroup: ({ index, chemical: chemicalName }) => {
+      creatorChemicals.removeChemicalFromSelectedGroup({ index, chemical: chemicalName })
+    }
   },
 }
 </script>
