+++ src/stores/creator-general.js
@@ -1,0 +1,46 @@
+import { getFormData, today, incrementField, resetCreator } from "@/lib/creator/creator-controls";
+
+import {defineStore} from "pinia";
+
+export const useCreatorGeneralStore = defineStore('creatorGeneral', {
+    state: () => ({
+        availableOrganisms: [],
+        availablePartners: [],
+        selectedOrganism: 1,
+        selectedPartner: null,
+        dates: [today.format(), today.format()],
+        solvent: 'WATER',
+        batch: 'AA',
+        controls: 4,
+        replicates: 4,
+        blanks: 3,
+        userOrganisation: null
+    }),
+    actions: {
+        setAvailableOrganisms(state, organisms) { state.availableOrganisms = organisms },
+        setAvailablePartners(state, partners) { state.availablePartners = partners },
+        setSelectedPartner(state, partner) { state.selectedPartner = partner },
+        setUserOrganisation(state, organisation) { state.userOrganisation = organisation },
+        setDates(state, dates) { state.dates = dates },
+        setSolvent(state, solvent) { state.solvent = solvent },
+        setSelectedOrganism(state, organism) { state.selectedOrganism = organism },
+        setBatch(state, batch) { state.batch = batch.toUpperCase() },
+        setControls(state, controls) { state.controls = controls },
+        setReplicates(state, replicates) { state.replicates = replicates },
+        setTimepoints(state, timepoints) { state.timepoints = timepoints },
+        setBlanks(state, blanks) { state.blanks = blanks },
+        backButton(state) {
+            state.created = false;
+            state.error = false;
+            state.loading = false
+        },
+        async getFormData({ commit }, token) { await getFormData(commit, token) },
+        changeField({ commit, state }, { field, value }) { incrementField(commit, state, field, value) },
+        resetForm({ commit, state }) { resetCreator(commit, state) },
+        sortDates({ commit }, dates) { commit('setDates', dates.sort((a, b) => new Date(a) - new Date(b)))},
+        getPartner: state => {
+            if (!state.selectedPartner) return 'UOB'
+            return state.availablePartners.find(partner => partner['name'] === state.selectedPartner)
+        }
+    }
+})
